+++ vscode-extension/src/extension.ts
@@ -10,6 +10,7 @@ import { locateCompiler, locateInterpreter, checkToolchain } from './locate';
 import { ensureUtf8Sources } from './convert';
 import { rewrapStringCommand, bindRewrapKeyCommand, registerRewrapAction } from './rewrap';
 import { registerEncodingFixes } from './quickfix';
+import { setupSpellChecking } from './spellcheck';
 import { startLanguageClient, stopLanguageClient, restartWhenServerSettingsChange,
     syncServerCompiler } from './client';
 import { initNotices, compilerNoticeSuppressed, suppressCompilerNotice } from './notices';
@@ -71,6 +72,7 @@ function registerCommands(context: ExtensionContext): void {
         commands.registerCommand('alanif.convertSources', () => ensureUtf8Sources()),
         commands.registerCommand('alanif.rewrapString', () => rewrapStringCommand()),
         commands.registerCommand('alanif.bindRewrapKey', () => bindRewrapKeyCommand()),
+        commands.registerCommand('alanif.setupSpellChecking', () => setupSpellChecking()),
     );
 }